MOved SetupConnection to public sql utils
