+++ server/channels/store/sqlstore/utils.go
@@ -5,7 +5,6 @@ package sqlstore
 
 import (
 	"database/sql"
-	"errors"
 	"io"
 	"net/url"
 	"strconv"
@@ -16,8 +15,6 @@ import (
 
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/shared/mlog"
-
-	"github.com/go-sql-driver/mysql"
 )
 
 var escapeLikeSearchChar = []string{
@@ -181,32 +178,6 @@ func DSNHasBinaryParam(dsn string) (bool, error) {
 // AppendBinaryFlag updates the byte slice to work using binary_parameters=yes.
 func AppendBinaryFlag(buf []byte) []byte {
 	return append([]byte{0x01}, buf...)
-}
-
-func SanitizeDataSource(driverName, dataSource string) (string, error) {
-	switch driverName {
-	case model.DatabaseDriverPostgres:
-		u, err := url.Parse(dataSource)
-		if err != nil {
-			return "", err
-		}
-		u.User = url.UserPassword("****", "****")
-		params := u.Query()
-		params.Del("user")
-		params.Del("password")
-		u.RawQuery = params.Encode()
-		return u.String(), nil
-	case model.DatabaseDriverMysql:
-		cfg, err := mysql.ParseDSN(dataSource)
-		if err != nil {
-			return "", err
-		}
-		cfg.User = "****"
-		cfg.Passwd = "****"
-		return cfg.FormatDSN(), nil
-	default:
-		return "", errors.New("invalid drivername. Not postgres or mysql.")
-	}
 }
 
 const maxTokenSize = 50
